requirements.txt
exifread
pandas
streamlit
folium
pillow_heif